Reorganize test structure by categories
- Created organized test directory structure:
  - tests/admin/ - Admin panel tests (resources, pages, widgets, actions, forms, tables)
  - tests/frontend/ - Frontend tests (pages, components, features, integration)
  - tests/api/ - API tests (endpoints, auth, validation)
  - tests/models/ - Model tests (core, business, relationships)
  - tests/services/ - Service tests (business, integration, external)
  - tests/browser/ - Browser/Dusk tests (admin, frontend, integration)

- Moved all test files to appropriate categories:
  - Admin Resource tests → tests/admin/resources/
  - Frontend Feature tests → tests/frontend/features/
  - Model tests → tests/models/ (core, business, relationships)
  - Browser tests → tests/browser/ (admin, frontend, integration)
  - API tests → tests/api/endpoints/

- Added comprehensive README.md documenting the new structure
- Removed old flat directory structure (tests/Feature/, tests/Unit/, tests/Browser/)
- Maintained all existing test functionality while improving organization

Benefits:
- Better organization and logical grouping
- Easier maintenance and navigation
- Clearer separation of concerns
- Improved team collaboration
- Scalable structure for future growth

+++ app/Filament/Widgets/DashboardOverviewWidget.php
@@ -70,7 +70,7 @@ class DashboardOverviewWidget extends BaseWidget
         // Advanced Metrics
         $avgOrderValue = $totalOrders > 0 ? $totalRevenue / $totalOrders : 0;
         $totalReviews = Review::where('is_approved', true)->count();
-        $avgRating = Review::where('is_approved', true)->avg('rating') ?? 0;
+        $avgRating = (float) (Review::where('is_approved', true)->avg('rating') ?? 0);
         $activeCampaigns = Campaign::where('status', 'active')->count();
         $totalCampaignViews = CampaignView::sum('views_count');
         $totalCampaignClicks = CampaignClick::sum('clicks_count');
